requirements.txt
# LangChain ecosystem
langchain==0.3.27
langchain-google-genai==2.0.10
langgraph==0.6.7
# Google Gemini
google-generativeai==0.8.5
fastapi==0.116.2
uvicorn==0.35.0
python-dotenv==1.0.1
python-multipart==0.0.20
requests==2.32.3
torch==2.8.0
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
elevenlabs
pyttsx3
pydub
langgraph 
google-generativeai 
python-dotenv 
requests
pillow